+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+##
+### Added
+- NullTime to handle NULL DATE columns
+
+### Changed
+- Return NullTime instead of time.Time for interface{} destination.
 
 ## [0.10.0]
 ### Added